order: update orders amount column to use numeric data type with precision of 6 and scale 22 to represent monetary value instead of real type that does not display nor have fixed decimals by default

+++ order/migrations/sqls/20200602055839-orders-up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS orders (
 	id UUID PRIMARY KEY,
-	amount REAL NOT NULL,
+	amount NUMERIC(6, 2) NOT NULL,
 	user_id UUID NOT NULL,
 	payee_id UUID NOT NULL,
 	description VARCHAR(500) NOT NULL,
